Extract and convert posterior parameters

This function extracts parameters from an object created by `rjuliabugs` and converts them to a specified posterior format.

Usage

get_params(rjuliabugs, params, posterior_type = "array")

Arguments

rjuliabugs

An object containing posterior samples and a `name` attribute used to extract parameters.

params

A character vector specifying which parameters to extract. **(NOTE: This argument is unused in the current function body.)**

posterior_type

A character string specifying the desired output format. Options are: "array", "rvar", "mcmc", "draws".

Value

The posterior samples converted to the specified format. The return type depends on posterior_type:

"array"

A 3D array of posterior samples.

"rvar"

An object of class rvar from the posterior package.

"mcmc"

An mcmc or mcmc.list object from the coda package.

"draws"

An object of class draws_array or similar from the posterior package.

Details

The function assumes that get_params_from_name() and other referenced variables such as params_to_save and n_chain are available in the current environment.

Examples

if (FALSE) { # \dontrun{
get_params(rjuliabugs = fit, params = c("alpha", "beta"), posterior_type = "array")
} # }